🐞 fix(rCore Tutorial): Stride Scheduling

+++ os/src/config.rs
@@ -16,9 +16,9 @@ pub const TRAP_CONTEXT: usize = TRAMPOLINE - PAGE_SIZE;
 
 pub const CLOCK_FREQ: usize = 12500000;
 
-pub const BIG_STRIDE: isize = isize::MAX;
+pub const BIG_STRIDE: usize = isize::MAX as usize;
 
-pub const DEFAULT_PRIO: isize = 16;
+pub const DEFAULT_PRIO: usize = 16;
 
 /// Return (bottom, top) of a kernel stack in kernel space.
 pub fn kernel_stack_position(app_id: usize) -> (usize, usize) {
